+++ src/app/utils/format.ts
@@ -24,3 +24,21 @@ export function convertTimeFormat(time) {
 export function padZero(num) {
   return num.toString().padStart(2, '0');
 }
+
+export const createFormData = async (url: string, fieldName: string): Promise<FormData | null> => {
+  try {
+    const response = await fetch(url);
+    if (!response.ok) {
+      throw new Error(`Failed to fetch the file from ${url}`);
+    }
+
+    const blob = await response.blob();
+    const formData = new FormData();
+    formData.append(fieldName, blob, 'tmp.mp4');
+
+    return formData;
+  } catch (error) {
+    console.error('Error creating FormData:', error);
+    return null;
+  }
+}
